killing method
Started work on making a method for killing the characters. Additionally, I began work on placing the newly initialised characters into the centre of the screen. This way the players will have a natural spawn point instead of being randomly placed. Additionally, I changed it so the characters will be scaled to adequate character scale when moved onto the main game, so i can continue to have big characters in the character creation screen.